% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event2car_range.R
\name{event2car_range}
\alias{event2car_range}
\title{Cumulative abnormal returns for firm(s) at event(s)}
\usage{
event2car_range(
  returns = NULL,
  regressor = NULL,
  event_date = NULL,
  method = c("mean_adj", "mrkt_adj_within", "mrkt_adj_out"),
  imputation_returns = c("approx", "mean", "drop", "pmm"),
  imputation_regressor = c("approx", "mean"),
  car_lag = 1,
  car_lead = 5,
  estimation_period = 150,
  range_before = 30,
  range_after = 30,
  output_format = c("zoo", "data.frame")
)
}
\arguments{
\item{returns}{an object of class \code{zoo} containing rates of returns of securities.}

\item{regressor}{an object of the same class as \code{returns} containing regressors.
The argument can be omitted, if market model is \code{mean_adj}.}

\item{event_date}{an character object or an object of class \code{Date} containing one event date or multiple event dates in the format YYYY-MM-DD.}

\item{method}{a character indicating the method used to calculate abnormal returns and cumulative abnormal returns. Choose among
\code{mean_adj}, \code{mrkt_adj_out}, and \code{mrkt_adj_within}.}

\item{imputation_returns}{a character indicating the way of dealing with missing values in returns data:
\code{drop}: No imputation at all and dropping observations with NAs,
\code{mean}: imputing missing data with the mean stock return (for further information see `?zoo::na.aggregate`) ,
\code{approx}: imputing missing data using interpolated values (for further information see `?zoo::na.approx`),
\code{pmm}: imputing missing data by dint of predictive mean matching used in the mice package (for further information see `?mice::mice`).}

\item{imputation_regressor}{a character indicating the way of dealing with missing values in regressor data:
\code{mean}: imputing missing data with the mean stock return (for further information see `?zoo::na.aggregate`) ,
\code{approx}: imputing missing data using interpolated values (for further information see `?zoo::na.approx`),}

\item{car_lag}{an object of class \code{intenger} measuring the start of the event window. The default is 1 day prior to the event date.}

\item{car_lead}{an object of class \code{intenger} measuring the end of the event window. The default is 5 days after the event date.}

\item{estimation_period}{an object of class \code{intenger} stating the number of days
prior to the event over which the market model parameters are estimated. Default is 150 days.
Note that the event period itself is not included in the event period to prevent the event from influencing the normal performance model parameter estimates.}

\item{range_before}{a numeric value indicating the time range to be included before the first event.}

\item{range_after}{a numeric value indicating the time range to be included after the last event.}

\item{output_format}{a character value signaling whether the cumulative abnormal returns
time series should be returned as \code{zoo} or \code{data.frame} object.}
}
\value{
an object of class \code{event2car_range} which contains the
cumulative abnormal returns time series (\code{car_timeseries}),
information on \code{event_dates},
information on \code{firms},
and information on \code{method}, \code{imputation_returns}, and \code{imputation_regressor}.
}
\description{
The function calculates cumulative abnormal returns
for a specific time-range before and after the event(s).
}
\details{
The package covers three models for the calculation of the cumulative abnormal returns:
- Mean-adjusted model (\code{mean_adj})
- Market-adjusted model using out-of-sample estimation (\emph{mrkt_adj_out})
- Market-adjusted model using within-sample estimation (\emph{mrkt_adj_within})
This is the logic suggested by multiple scholars. See references below.

The generic function is dispatched for such classes as
 \code{zoo}. (future versions of the package allow for classes of \code{data.frame}.)

If \code{method} is \emph{mrkt_adj_out} or \code{mrkt_adj_within}
and \code{regressor} has the length greater than one, the first element of
\code{regressor} will be applied for each security in \code{returns}.
}
\examples{
trumpelection <- as.Date("2016-11-08")
returns_firms <- tech_returns[,2:19]
return_indx <- tech_returns[,1]
effect_trump <- event2car_range(returns=returns_firms,regressor=return_indx,
                                imputation_returns="mean",
                                event_date=trumpelection,method="mean_adj")
}
\references{
MacKinlay, A.C. \emph{Event Studies in Economics and Finance}.
Journal of Economic Literature, 35(1):13-39, 1997.

Brown S.J., Warner J.B. \emph{Using Daily Stock Returns, The Case
of Event Studies}. Journal of Financial Economics, 14:3-31, 1985.

Davies, R., Studnicka, Z. \emph{The heterogeneous impact of Brexit: Early indications from the FTSE}.
European Economic Review, 110:1-17, 2018.
}
\keyword{eventstudy}
\keyword{finance}
\keyword{stock}
\keyword{timeseries}
