% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.event2car_range}
\alias{plot.event2car_range}
\title{Plots cumulative abnormal returns over time}
\usage{
\method{plot}{event2car_range}(
  x,
  ...,
  event_color = "red",
  event_linetype = "twodash",
  event_size = 1,
  event_label_size = 3,
  line_color = "black",
  background_color = "gray",
  background_alpha = 0.5,
  point_size = 0.2,
  point_alpha = 0.1,
  axis_text_size = 8,
  axis_title_size = 10,
  show_method = TRUE
)
}
\arguments{
\item{x}{an object of class \code{event2car_range}.}

\item{...}{other arguments ignored (for compatibility with generic)}

\item{event_color}{a character object indicating the color of the vertical lines at the \code{event_date}. Default is "red".}

\item{event_linetype}{a character object indicating the linetype of the veritical line at the \code{event_date}. Default is "twodash".}

\item{event_size}{a numeric value indicating the size of the vertical line at the \code{event_date}. Default is 1.}

\item{event_label_size}{a numeric value indicating the size of the event label(s). Default is 3.}

\item{line_color}{a character object indicating the color of the loess curve (see \code{stats::loess}) of all
cumulative abnormal returns. Default is "black".}

\item{background_color}{a character value indicating the color of the area above zero. Default is "gray".}

\item{background_alpha}{a numeric value indicating the transparency of the area above zero. Default is 0.5.}

\item{point_size}{a numeric value specifying the size of the points representing
cumulative abnormal returns. Default is 0.2.}

\item{point_alpha}{a numeric value specifying the transparency of the points
representing cumulative abnormal returns. Default is 0.1.}

\item{axis_text_size}{a numeric value indicating the size of the axis text. Default is 8.}

\item{axis_title_size}{a numeric value indicating the size of the axis titles. Default is 10.}

\item{show_method}{a logical value indicating whether the information on the applied method
(\code{mean_adj}, \code{mrkt_adj_out}, or \code{mrkt_adj_within}) should be included in the plot.}
}
\description{
plot method for the class "event2car_range".
}
\examples{
trumpelection <- as.Date("2016-11-08")
returns_firms <- tech_returns[,2:19]
return_indx <- tech_returns[,1]
effect_trump <- event2car_range(returns=returns_firms,regressor=return_indx,
                                imputation_returns="mean",
                                event_date=trumpelection,method="mean_adj")
plot(effect_trump)
}
\keyword{internal}
