% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event2car.R
\name{event2car}
\alias{event2car}
\title{event2car: Calculates \code{CAR} (cumulative abnormal returns) for firm(s) and event(s)}
\usage{
event2car(
  returns = NULL,
  regressor = NULL,
  event_date = NULL,
  method = c("mean_adj", "mrkt_adj_within", "mrkt_adj_out"),
  imputation = c("mean", "none", "pmm"),
  car_lag = 1,
  car_lead = 5,
  estimation_period = 150
)
}
\arguments{
\item{returns}{an object of \code{data.frame} or \code{zoo} containing rates of returns of securities.}

\item{regressor}{an object of the same class as \code{returns} containing regressors.
The argument can be omitted, if market model is \code{mean_adj}.}

\item{event_date}{an character object or an object of class \code{Date} containing one event date or multiple event dates in the format YYYY-MM-DD.}

\item{method}{a character indicating the method used to calculate abnormal returns and cumulative abnormal returns. Choose among
\code{mean_adj}, \code{mrkt_adj_out}, and \code{mrkt_adj_within}.}

\item{imputation}{a character indicating the way of dealing with missing values:
No imputation at all and dropping observations with NAs (\code{none}),
imputing missing data with the mean stock return (\code{mean}),
imputing missing data by dint of predictive mean matching used in the mice package (\code{pmm}).}

\item{car_lag}{an object of class \code{intenger} measuring the start of the event window. The default is 1 day prior to the event date.}

\item{car_lead}{an object of class \code{intenger} measuring the end of the event window. The default is 5 days after the event date.}

\item{estimation_period}{an object of class \code{intenger} stating the number of days
prior to the event over which the market model parameters are estimated. Default is 150 days.
Note that the event period itself is not included in the event period to prevent the event from influencing the normal performance model parameter estimates.}
}
\value{
an object of class \code{list} which contains abnormal returns on the event date(s),
confidence intervals of the abnormal returns, and the cumulative abnormal return of the event period(s).
Note that the \code{mean_adj} adn the \code{mrkt_adj_out} method produce mean abnormal returns,
yet the \code{mrkt_adj_within} method delivers coefficients of the event period.
}
\description{
The function calculates abnormal returns and confidence intervals
during the event period(s) as well as
cumulative abnormal returns of event(s).
}
\details{
The package covers three models for the calculation of the cumulative abnormal returns:
- Mean-adjusted model (\code{mean_adj})
- Market-adjusted model using out-of-sample estimation (\emph{mrkt_adj_out})
- Market-adjusted model using within-sample estimation (\emph{mrkt_adj_within})
This is the logic suggested by multiple scholars. See references below.

The generic function is dispatched for such classes as
 \code{zoo}. (future versions of the package allow for classes of \code{data.frame}.)

If \code{market_model} is \emph{mrkt_adj_out} or \code{mrkt_adj_within}
and \code{regressor} has the length greater than one, the first element of
\code{regressor} will be applied for each security in \code{returns}.
}
\examples{
# prepare data
trumpelection <- as.Date("2016-11-08")
returns_firms <- tech_returns[,2:19]
return_indx <- tech_returns[,1]
# mean adjusted model
event2car(returns=returns_firms,regressor=return_indx,
          event_date=trumpelection,method="mean_adj")
# market adjusted model (out-of sample estimation)
event2car(returns=returns_firms,regressor=return_indx,
          event_date=trumpelection,method="mrkt_adj_out")
# market adjusted model (within sample estimation)
event2car(returns=returns_firms,regressor=return_indx,
          event_date=trumpelection,method="mrkt_adj_within")
}
\references{
MacKinlay, A.C. \emph{Event Studies in Economics and Finance}.
Journal of Economic Literature, 35(1):13-39, 1997.

Brown S.J., Warner J.B. \emph{Using Daily Stock Returns, The Case
of Event Studies}. Journal of Financial Economics, 14:3-31, 1985.

Davies, R., Studnicka, Z. \emph{The heterogeneous impact of Brexit: Early indications from the FTSE}.
European Economic Review, 110:1-17, 2018.
}
\keyword{eventstudy}
\keyword{finance}
\keyword{stock}
